betty style

+++ hello_world/6-size.c
@@ -7,10 +7,10 @@
  */
 int main(void)
 {
-printf("Size of a char: %lu bytes(s)\n", sizeof(char));
-printf("Size of int: %lu bytes(s)\n", sizeof(int));
-printf("Size of long int: %lu bytes(s)\n", sizeof(long int));
-printf("Size of long long int: %lu bytes(s)\n", sizeof(long long int));
+printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(char));
+printf("Size of int: %lu bytes(s)\n", (unsigned long)sizeof(int));
+printf("Size of long int: %lu bytes(s)\n", (unsigned long)sizeof(long int));
+printf("Size of long long int: %lu bytes(s)\n", (unsigned long)sizeof(long long int));
 
 return (0);
 }
